Add option to include a direct feedback file link
This means that if the progress.json file has achieved points and a feedback dile the achieved points string will display the same but also be a link to the feedback file.
The feedback file is relative to the submission directory (if given - otherwise it is  relative to the README).

Dependencies were updated during this modification.